Add maximum ride price to Ride window
A new "Maximum Ride Price" row is added to the Ride window,
displaying the maximum price that can be set for said ride
before no peeps will decide to pay for the ride anymore.
This is equal to two times the ride value.

This figure is currently often calculated using external
tools or managed using a plugin that sets ride prices.

A new string, STR_6576, is added for display of the maximum
price. It has been added to the en-GB and en-US localisations.